Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
PyJWT==2.8.0